requirements.txt
beir
faiss-cpu
numba
numpy
pybind11
pytrec-eval
scikit-learn
torch
tqdm
